move interface and type to folder by feature
