+++ src/lib/client/UI/Search/SearchFeature/Interface.ts
@@ -1,15 +1,15 @@
-import type { SearchType } from '$lib/client/Utils/types';
+export type SearchType = 'none' | 'fuzzy' | 'detail';
+
+export interface SearchConditions {
+	query: string | null;
+	bookTitle: string | null;
+	author: string | null;
+	isbn: string | null;
+}
 
 export interface SearchProps {
 	searchType: SearchType;
 	searchConditions: SearchConditions;
 	pageCount: number;
 	startIndex: number;
-}
-
-export interface SearchConditions {
-	query: string;
-	bookTitle: string;
-	author: string;
-	isbn: string;
 }
